backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
hypothesis==6.92.1
groq==1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pdfplumber>=0.10.0
numpy>=1.24.0
python-docx>=1.0.0
pytesseract>=0.3.10
python-telegram-bot>=21.0
